fix(@clayui/color-picker): Move Hue input group to Hue internal component and SF
